+++ src/actions/index.js
@@ -1,12 +1,12 @@
-const CREATE_BOOK = 'CREATE_BOOK';
-const REMOVE_BOOK = 'REMOVE_BOOK';
+export const CREATE_BOOK = 'CREATE_BOOK';
+export const REMOVE_BOOK = 'REMOVE_BOOK';
 
-const createBook = ({ book }) => ({
+export const createBook = book => ({
   type: CREATE_BOOK,
   book,
 });
 
-const removeBook = ({ book }) => ({
+export const removeBook = book => ({
   type: REMOVE_BOOK,
   book,
 });