+++ src/reducers/books.js
@@ -1,14 +1,16 @@
-const CREATE_BOOK = 'CREATE_BOOK';
-const REMOVE_BOOK = 'REMOVE_BOOK';
+import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
 
 const booksReducer = (state = [], action) => {
   switch (action.type) {
     case CREATE_BOOK:
-      return [...state, action.book];
+      return { books: [...state.books, action.book] };
     case REMOVE_BOOK:
-      return [
-        ...state.slice(0, action.book.index),
-        ...state.slice(action.book.index + 1, state.length)];
+      return {
+        books:
+        [...state.books.slice(0, action.book),
+          ...state.books.slice(action.book + 1,
+            state.books.length)],
+      };
     default:
       return state;
   }