+++ src/reducers/index.js
@@ -1,4 +1,4 @@
 import { combineReducers } from 'redux';
 import booksReducer from './books';
 
-export default combineReducers({ booksReducer });
+export default combineReducers({ books: booksReducer });